+++ Assets/Textures/Content/Projectiles/AbsorbingLiquidEffect.png
