+++ Assets/Textures/Content/Projectiles/CEFist.png
