+++ Assets/Textures/Content/Projectiles/CursedBloodEffect.png
