+++ Assets/Textures/Content/Projectiles/FlyingKunai.png
