+++ Assets/Textures/Content/Projectiles/LightBeam.png
