+++ Assets/Textures/Content/Projectiles/LightCone.png
