+++ Assets/Textures/Content/Projectiles/SlowingAura.png
