+++ Content/Projectiles/AbsorbingLiquidEffect.cs
@@ -1,6 +1,6 @@
+using Microsoft.Xna.Framework.Graphics;
 using System;
 using System.Collections.Generic;
-using System.Diagnostics.Metrics;
 using Terraria;
 using Terraria.GameContent;
 using Terraria.ModLoader;
@@ -9,6 +9,8 @@ namespace TheBindingOfRarria.Content.Projectiles;
 
 public class AbsorbingLiquidEffect : ModProjectile
 {
+    public override string Texture => ContentPath + "Projectiles/" + Name;
+
     public override void SetStaticDefaults() => Main.projFrames[Type] = 4;
 
     public override void SetDefaults()
@@ -22,36 +24,31 @@ public class AbsorbingLiquidEffect : ModProjectile
         Projectile.netImportant = true;
         Projectile.timeLeft = 36;
     }
-    public override void AI()
-    {
 
-    }
     public override void OnKill(int timeLeft)
     {
-        var owner = Main.player[Projectile.owner];
+        Player owner = Main.player[Projectile.owner];
 
         if (owner.active)
             owner.Heal((int)Projectile.ai[0]);
-
-        base.OnKill(timeLeft);
     }
+
     public override void DrawBehind(int index, List<int> behindNPCsAndTiles, List<int> behindNPCs, List<int> behindProjectiles, List<int> overPlayers, List<int> overWiresUI)
     {
-        base.DrawBehind(index, behindNPCsAndTiles, behindNPCs, behindProjectiles, overPlayers, overWiresUI);
         Main.instance.DrawCacheNPCsOverPlayers.Add(index);
         overPlayers.Add(index);
     }
+
     public override bool PreDraw(ref Color lightColor)
     {
-        var texture = TextureAssets.Projectile[Type].Value;
+        Texture2D texture = TextureAssets.Projectile[Type].Value;
 
         int frame = Math.Max(0, 8 - Projectile.timeLeft / 4);
-        var rect = texture.Frame(1, 9, 0, frame, 0, -2);
+        Rectangle rect = texture.Frame(1, 9, 0, frame, 0, -2);
 
+        Color color = Color.White;
 
-        var color = Color.White;
-
-        Main.spriteBatch.Draw(texture, Projectile.Center - Main.screenPosition - Projectile.Size / 2, rect, color);
+        Main.spriteBatch.Draw(texture, Projectile.Center - Main.screenPosition - Projectile.Size * 0.5f, rect, color);
 
         return false;
     }
